# codingTest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "All",
                "description": "All",
                "responses": {
                    "default": {
                        "description": "All Data"
                    }
                }
            }
        },
        "/candidate": {
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Add",
                "description": "Add",
                "operationId": "9ecb221cf433b8e349f9c42fed065521",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Add Candidate",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Julle"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Show page"
                    }
                }
            }
        },
        "/candidate/{id}": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Show",
                "description": "Show",
                "operationId": "e78919f034bb69d4ba7d00b37ca7c327",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "candidate id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Show page"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}